More CR comments
